Nested optimization solution for LQR
This file contains an adjoint-based nested-optimization solution for imitation learning of an LQR-governed system. Currently, the code is not polished, and there are many areas that where heavy computations are redundantly being done. The code must also be generalized in order to cover any system a user wants, instead of just functioning as an LQR learner. 

Signed-off-by: apandey02 <93224425+apandey02@users.noreply.github.com>